+++ client/shared/src/search/query/diagnostics.test.ts
@@ -19,22 +19,19 @@ describe('getDiagnostics()', () => {
     })
 
     test('invalid filter value', () => {
-        expect(
-            getDiagnostics(toSuccess(scanSearchQuery('case:maybe')), SearchPatternType.literal)
-        ).toMatchInlineSnapshot(
-            `
+        expect(getDiagnostics(toSuccess(scanSearchQuery('case:maybe')), SearchPatternType.literal))
+            .toMatchInlineSnapshot(`
             [
               {
                 "severity": 8,
-                "message": "Invalid filter value, expected one of: yes, no.",
+                "message": "Error: Invalid filter value, expected one of: yes, no.",
                 "startLineNumber": 1,
                 "endLineNumber": 1,
                 "startColumn": 1,
                 "endColumn": 5
               }
             ]
-        `
-        )
+        `)
     })
 
     test('search query containing colon, literal pattern type, do not raise error', () => {
@@ -53,5 +50,37 @@ describe('getDiagnostics()', () => {
         expect(
             getDiagnostics(toSuccess(scanSearchQuery('repo:a (file:b and c)')), SearchPatternType.regexp)
         ).toMatchInlineSnapshot('[]')
+    })
+
+    test('search query with empty filter diagnostic', () => {
+        expect(getDiagnostics(toSuccess(scanSearchQuery('meatadata file: ')), SearchPatternType.regexp))
+            .toMatchInlineSnapshot(`
+            [
+              {
+                "severity": 4,
+                "message": "Warning: This filter is empty. Remove the space between the filter and value or quote the value to include the space. E.g., file:\\" a term\\".",
+                "startLineNumber": 1,
+                "endLineNumber": 1,
+                "startColumn": 11,
+                "endColumn": 15
+              }
+            ]
+        `)
+    })
+
+    test('search query with both invalid filter and empty filter returns only one diagnostic for the first issue', () => {
+        expect(getDiagnostics(toSuccess(scanSearchQuery('meatadata type: ')), SearchPatternType.regexp))
+            .toMatchInlineSnapshot(`
+            [
+              {
+                "severity": 8,
+                "message": "Error: Invalid filter value, expected one of: diff, commit, symbol, repo, path, file.",
+                "startLineNumber": 1,
+                "endLineNumber": 1,
+                "startColumn": 11,
+                "endColumn": 15
+              }
+            ]
+        `)
     })
 })